requirements.txt
annotated-types==0.6.0
anyio==4.3.0
argcomplete==1.10.3
backports.zoneinfo==0.2.1
beautifulsoup4==4.8.2
bitarray==2.9.2
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
compressed_rtf==1.0.6
docx==0.2.4
docx2txt==0.8
ebcdic==1.1.1
exceptiongroup==1.2.1
extract-msg==0.28.7
fastapi==0.110.2
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.7
IMAPClient==2.1.0
lxml==5.2.1
olefile==0.47
pdfminer.six==20191110
pdfreader==0.1.12
pillow==10.3.0
plum-dispatch==1.7.4
pycryptodome==3.20.0
pydantic==2.7.0
pydantic_core==2.18.1
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
python-multipart==0.0.9
python-pptx==0.6.23
PyYAML==6.0.1
requests==2.31.0
six==1.12.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.8.1
Spire.Doc==12.3.2
SQLAlchemy==2.0.29
starlette==0.37.2
textract==1.6.5
tika==2.6.0
typing_extensions==4.11.0
tzlocal==5.2
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
xlrd==1.2.0
XlsxWriter==3.2.0
